fix: move check on env variable CONTAINER_REGISTRY_BUCKET inside check_prerequisites plus other little fixes
Also remove `-u` from the `set` directive: we're checking for all the variables at the start of the install or uninstall functions, thus we do not need to enforce bash to imm. terminate the execution in case of unbound variables (which in turn does not permit easily default values for env variables).

Also fixing an invokation of AWS_CMD

Signed-off-by: Leonardo Di Donato <leodidonato@gmail.com>