Remove Ingress provision and related things.
This PR removes the Ingress (and thus ALB) provision code from the
guide to let users do it based on their use-case i.e how they
want to expose the `proxy` svc to the internet with SSL.

In the easiest path, they can just use the `proxy` service's CLB URL
(created by default), and use it everywhere and things would work.
(We'll be using whatever `certificate` config option has, as the SSL
cert on the proxy)

This PR also adds a note on how to use SSH gateway stuff when a ALB
is used.

Signed-off-by: Tarun Pothulapati <tarun@gitpod.io>